+++ src/migrations/0003_users_table.py
@@ -1,14 +1,15 @@
+# pylint: disable=C0103,R0914
 """
-Add additional column 'created_at' and replace column 'timestamp' with 'updated_at' in the messages table.
-https://github.com/obervinov/pyinstabot-downloader/issues/62
+Add additional column 'status' in the users table.
+https://github.com/obervinov/users-package/blob/v3.0.0/tests/postgres/tables.sql
 """
 VERSION = '1.0'
-NAME = '0002_messages_table'
+NAME = '0003_users_table'
 
 
 def execute(obj):
     """
-    Add additional column 'created_at' and replace column 'timestamp' with 'updated_at' in the messages table.
+    Add additional column 'status' in the users table.
 
     Args:
         obj: An obj containing the database connection and cursor, as well as the Vault instance.
@@ -17,14 +18,13 @@ def execute(obj):
         None
     """
     # database settings
-    table_name = 'messages'
-    rename_columns = [('timestamp', 'updated_at')]
-    add_columns = [('created_at', 'TIMESTAMP', 'CURRENT_TIMESTAMP'), ('state', 'VARCHAR(255)', "'added'")]
-    print(f"{NAME}: Start migration for the {table_name} table: Rename columns {rename_columns}, Add columns {add_columns}...")
+    table_name = 'users'
+    add_columns = [('status', 'VARCHAR(255)', "'denied'")]
+    print(f"{NAME}: Start migration for the {table_name} table: Add columns {add_columns}...")
 
+    # check if the table exists and has the necessary schema for execute the migration
     conn = obj.get_connection()
     with conn.cursor() as cursor:
-        # check if the table exists and has the necessary schema for execute the migration
         # check table
         cursor.execute("SELECT * FROM information_schema.tables WHERE table_schema = 'public' AND table_name = %s;", (table_name,))
         table = cursor.fetchone()
@@ -35,22 +35,11 @@ def execute(obj):
 
         if not table:
             print(f"{NAME}: The {table_name} table does not exist. Skip the migration.")
+
         elif len(columns) < 1:
             print(f"{NAME}: The {table_name} table does not have the necessary columns to execute the migration. Skip the migration.")
-        else:
-            for column in rename_columns:
-                try:
-                    print(f"{NAME}: Rename column {column[0]} to {column[1]} in the {table_name} table...")
-                    cursor.execute(f"ALTER TABLE {table_name} RENAME COLUMN {column[0]} TO {column[1]}")
-                    conn.commit()
-                    print(f"{NAME}: Column {column[0]} has been renamed to {column[1]} in the {table_name} table.")
-                except obj.errors.DuplicateColumn as error:
-                    print(f"{NAME}: Columns in the {table_name} table have already been renamed. Skip renaming: {error}")
-                    conn.rollback()
-                except obj.errors.UndefinedColumn as error:
-                    print(f"{NAME}: Columns in the {table_name} table have not been renamed. Skip renaming: {error}")
-                    conn.rollback()
 
+        else:
             for column in add_columns:
                 try:
                     print(f"{NAME}: Add column {column[0]} to the {table_name} table...")
@@ -63,4 +52,3 @@ def execute(obj):
                 except obj.errors.FeatureNotSupported as error:
                     print(f"{NAME}: Columns in the {table_name} table have not been added. Skip adding: {error}")
                     conn.rollback()
-    obj.close_connection(conn)