+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## v2.3.0 - 2024-08-25
+### What's Changed
+**Full Changelog**: https://github.com/obervinov/pyinstabot-downloader/compare/v2.2.1...v2.3.0 by @obervinov in https://github.com/obervinov/pyinstabot-downloader/pull/95
+#### 💥 Breaking Changes
+* now all user data is stored in the database
+* psql credentials are now written out via Vault Database Engine
+#### 🚀 Features
+* bump vault-package to major version `3.0.0`
+* bump users-package to major version `3.0.0`
+* bump telegram-package to major version `2.0.1`
+* [Switch reading of the database connection configuration to db engine](https://github.com/obervinov/pyinstabot-downloader/issues/33)
+
+
 ## v2.2.1 - 2024-08-24
 ### What's Changed
 **Full Changelog**: https://github.com/obervinov/pyinstabot-downloader/compare/v2.2.0...v2.2.1 by @obervinov in https://github.com/obervinov/pyinstabot-downloader/pull/94